02-ai-agents/05-crash-course/openai-docs-practice/05-repl-utility/pyproject.toml
[project]
name = "05-repl-utility"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
